+++ core/mod.ts
@@ -1,9 +1,9 @@
 export * from "./authentication.ts";
 export * from "./authorization.ts";
 export * from "./configuration.ts";
+export * from "./cors.ts";
 export * from "./fetch-router.ts";
 export * from "./http.ts";
-export * from "./tokens.ts";
 export * from "./migrator.ts";
 export * from "./postgres.ts";
 export * from "./random.ts";
@@ -12,4 +12,5 @@ export * from "./store.ts";
 export * from "./structures.ts";
 export * from "./terminator.ts";
 export * from "./timers.ts";
+export * from "./tokens.ts";
 export * from "./utilities.ts";